# backend/openapi.yaml
openapi: 3.0.0
paths:
...
  /api/v1/admin/dogs:
    get:
      operationId: DogsAdminController_findAll
      parameters:
        - name: page
          required: false
          in: query
          schema:
            default: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            default: 20
            type: number
        - name: status
          required: false
          in: query
          schema:
            type: string
            enum:
              - AVAILABLE
              - ADOPTED
              - IN_CARE
              - ARCHIVED
        - name: isPublished
          required: false
          in: query
          schema:
            type: boolean
        - name: excludeArchived
          required: false
          in: query
          description: When true, omit ARCHIVED dogs (unless status=ARCHIVED)
          schema:
            type: boolean
        - name: search
          required: false
          in: query
          description: Filter by slug substring (case-insensitive)
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedDogsAdminResponseDto'
      security: &ref_3
        - bearer: []
      summary: List dogs (admin)
      tags: &ref_4
        - admin-dogs
    post:
      operationId: DogsAdminController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDogDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DogAdminResponseDto'
      security: *ref_3
      summary: Create dog
      tags: *ref_4
  /api/v1/admin/dogs/{id}:
    get:
      operationId: DogsAdminController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DogAdminResponseDto'
      security: *ref_3
      summary: Get dog by id (admin)
      tags: *ref_4
    patch:
      operationId: DogsAdminController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDogDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DogAdminResponseDto'
      security: *ref_3
      summary: Update dog
      tags: *ref_4
  /api/v1/admin/dogs/{id}/archive:
    post:
      operationId: DogsAdminController_archive
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DogAdminResponseDto'
      security: *ref_3
      summary: Archive dog (ADMIN only)
      tags: *ref_4
  /api/v1/dogs:
    get:
      operationId: DogsPublicController_findAll
      parameters:
        - name: page
          required: false
          in: query
          schema:
            default: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            default: 20
            type: number
        - name: status
          required: false
          in: query
          schema:
            type: string
            enum:
              - AVAILABLE
              - ADOPTED
              - IN_CARE
              - ARCHIVED
        - name: isPublished
          required: false
          in: query
          schema:
            type: boolean
        - name: excludeArchived
          required: false
          in: query
          description: When true, omit ARCHIVED dogs (unless status=ARCHIVED)
          schema:
            type: boolean
        - name: search
          required: false
          in: query
          description: Filter by slug substring (case-insensitive)
          schema:
            type: string
        - name: accept-language
          required: true
          in: header
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedDogsPublicResponseDto'
      summary: List published dogs
      tags: &ref_5
        - dogs
  /api/v1/dogs/{slug}:
    get:
      operationId: DogsPublicController_findBySlug
      parameters:
        - name: slug
          required: true
          in: path
          schema:
            type: string
        - name: accept-language
          required: true
          in: header
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DogPublicResponseDto'
      summary: Get published dog by slug
      tags: *ref_5
...
  /api/v1/admin/crypto-addresses:
    get:
      operationId: CryptoAddressesAdminController_findAll
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CryptoAddressesAdminResponseDto'
      security: &ref_13
        - bearer: []
      summary: List crypto donation addresses
      tags: &ref_14
        - admin-crypto-addresses
    post:
      operationId: CryptoAddressesAdminController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCryptoAddressDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CryptoAddressAdminDto'
      security: *ref_13
      summary: Create crypto donation address
      tags: *ref_14
  /api/v1/admin/crypto-addresses/{id}:
    patch:
      operationId: CryptoAddressesAdminController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCryptoAddressDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CryptoAddressAdminDto'
      security: *ref_13
      summary: Update crypto donation address
      tags: *ref_14
  /api/v1/admin/donations:
    get:
      operationId: DonationsAdminController_findAll
      parameters:
        - name: page
          required: false
          in: query
          schema:
            default: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            default: 20
            type: number
        - name: status
          required: false
          in: query
          schema:
            type: string
            enum:
              - PENDING
              - CONFIRMED
              - FAILED
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedDonationsResponseDto'
      security: &ref_15
        - bearer: []
      summary: List donation records
      tags: &ref_16
        - admin-donations
    post:
      operationId: DonationsAdminController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDonationAdminDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DonationResponseDto'
      security: *ref_15
      summary: Create donation record manually
      tags: *ref_16
  /api/v1/admin/donations/{id}:
    patch:
      operationId: DonationsAdminController_updateStatus
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDonationStatusDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DonationResponseDto'
      security: *ref_15
      summary: Confirm or reject pending donation
      tags: *ref_16
...
  /api/v1/admin/content/pages:
    get:
      operationId: ContentAdminController_listPages
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContentPageSummaryDto'
      security: &ref_19
        - bearer: []
      summary: List CMS pages available for editing
      tags: &ref_20
        - admin-content
  /api/v1/admin/content/pages/{entityId}:
    get:
      operationId: ContentAdminController_getPage
      parameters:
        - name: entityId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageContentAdminDto'
      security: *ref_19
      summary: Get all stored translations for a page
      tags: *ref_20
    put:
      operationId: ContentAdminController_upsertPage
      parameters:
        - name: entityId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertPageContentDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageContentAdminDto'
      security: *ref_19
      summary: Bulk upsert page translations
      tags: *ref_20
info:
  title: Dog Rescue API
  description: REST API for Dog Rescue (MVP scaffold)
  version: '1.0'
  contact: {}
...
servers:
  - url: /
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    MediaResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        mimeType:
          type: string
          example: image/png
        sizeBytes:
          type: number
          example: 1024
        url:
          type: string
          description: Presigned GET URL (expires per MEDIA_PRESIGNED_TTL_SECONDS)
        createdAt:
          type: string
          example: '2026-08-02T00:00:00.000Z'
      required:
        - id
        - mimeType
        - sizeBytes
        - url
        - createdAt
    DogListItemDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        status:
          type: string
          enum:
            - AVAILABLE
            - ADOPTED
            - IN_CARE
            - ARCHIVED
        isPublished:
          type: boolean
        nameEn:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - slug
        - status
        - isPublished
        - updatedAt
    PaginatedDogsAdminResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DogListItemDto'
        total:
          type: number
        page:
          type: number
        limit:
          type: number
      required:
        - items
        - total
        - page
        - limit
    DogLocaleContentRequiredDto:
      type: object
      properties:
        name:
          type: string
          example: Luna
        description:
          type: string
          example: Friendly mixed-breed looking for a calm home.
        rescueStory:
          type: string
      required:
        - name
        - description
    DogLocaleContentOptionalDto:
      type: object
      properties:
        name:
          type: string
          example: Luna
        description:
          type: string
        rescueStory:
          type: string
    DogDescriptionsDto:
      type: object
      properties:
        en:
          $ref: '#/components/schemas/DogLocaleContentRequiredDto'
        th:
          $ref: '#/components/schemas/DogLocaleContentOptionalDto'
        ru:
          $ref: '#/components/schemas/DogLocaleContentOptionalDto'
      required:
        - en
    CreateDogDto:
      type: object
      properties:
        descriptions:
          $ref: '#/components/schemas/DogDescriptionsDto'
        slug:
          type: string
          example: luna
        status:
          type: string
          enum:
            - AVAILABLE
            - ADOPTED
            - IN_CARE
            - ARCHIVED
          default: IN_CARE
        isPublished:
          type: boolean
          default: false
      required:
        - descriptions
    DogAdminResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        status:
          type: string
          enum:
            - AVAILABLE
            - ADOPTED
            - IN_CARE
            - ARCHIVED
        isPublished:
          type: boolean
        descriptions:
          type: object
        seo:
          type: object
        media:
          type: array
          items:
            $ref: '#/components/schemas/MediaResponseDto'
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - slug
        - status
        - isPublished
        - descriptions
        - seo
        - media
        - createdAt
        - updatedAt
    UpdateDogDto:
      type: object
      properties:
        descriptions:
          $ref: '#/components/schemas/DogDescriptionsDto'
        seo:
          type: object
        slug:
          type: string
        status:
          type: string
          enum:
            - AVAILABLE
            - ADOPTED
            - IN_CARE
            - ARCHIVED
        isPublished:
          type: boolean
    DogPublicMediaDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
        mimeType:
          type: string
      required:
        - id
        - url
        - mimeType
    DogPublicResponseDto:
      type: object
      properties:
        slug:
          type: string
        status:
          type: string
          enum:
            - AVAILABLE
            - ADOPTED
            - IN_CARE
            - ARCHIVED
        name:
          type: string
        description:
          type: string
        rescueStory:
          type: string
        seoTitle:
          type: string
        seoDescription:
          type: string
        media:
          type: array
          items:
            $ref: '#/components/schemas/DogPublicMediaDto'
        locale:
          type: string
      required:
        - slug
        - status
        - name
        - description
        - media
        - locale
    PaginatedDogsPublicResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DogPublicResponseDto'
        total:
          type: number
        page:
          type: number
        limit:
          type: number
      required:
        - items
        - total
        - page
        - limit
...
    DonationResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        amount:
          type: string
          example: '1000.00'
        currency:
          type: string
          example: THB
        status:
          type: string
          enum:
            - PENDING
            - CONFIRMED
            - FAILED
        paymentMethod:
          type: string
          enum:
            - BANK
            - CRYPTO
        donorName:
          type: string
          nullable: true
        donorEmail:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - amount
        - currency
        - status
        - createdAt
        - updatedAt
    CryptoAddressAdminDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        currencyCode:
          type: string
          example: BTC
        label:
          type: object
          example: Bitcoin
        address:
          type: string
        isActive:
          type: boolean
        isDisplayed:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - currencyCode
        - address
        - isActive
        - isDisplayed
        - createdAt
        - updatedAt
    CryptoAddressesAdminResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CryptoAddressAdminDto'
      required:
        - items
    CreateCryptoAddressDto:
      type: object
      properties:
        currencyCode:
          type: string
          example: BTC
          description: Currency ticker (2-10 alphanumeric)
        label:
          type: string
          example: Bitcoin
        address:
          type: string
          example: bc1qexampleaddressfordevonly000000000
        setAsDisplayed:
          type: boolean
          description: Show this address on the public donate page (replaces current displayed address for this currency)
      required:
        - currencyCode
        - address
    UpdateCryptoAddressDto:
      type: object
      properties:
        address:
          type: string
        label:
          type: string
        isActive:
          type: boolean
        isDisplayed:
          type: boolean
          description: Show on public donate page (only one per currency)
    PaginatedDonationsResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DonationResponseDto'
        total:
          type: number
        page:
          type: number
        limit:
          type: number
      required:
        - items
        - total
        - page
        - limit
    CreateDonationAdminDto:
      type: object
      properties:
        amount:
          type: number
          example: 1000
        donorName:
          type: string
          example: Somchai
        donorEmail:
          type: string
          example: donor@example.com
        paymentMethod:
          type: string
          enum:
            - BANK
            - CRYPTO
        currency:
          type: string
          default: THB
        status:
          type: string
          enum:
            - PENDING
            - CONFIRMED
            - FAILED
          default: PENDING
      required:
        - amount
    UpdateDonationStatusDto:
      type: object
      properties:
        status:
          type: string
          enum:
            - CONFIRMED
            - FAILED
      required:
        - status
...
    ContentPageSummaryDto:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        fields:
          type: array
          items:
            type: string
      required:
        - id
        - label
        - fields
    ContentItemDto:
      type: object
      properties:
        locale:
          type: string
          enum:
            - en
            - th
            - ru
        field:
          type: string
        value:
          type: string
      required:
        - locale
        - field
        - value
    PageContentAdminDto:
      type: object
      properties:
        entityId:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/ContentItemDto'
      required:
        - entityId
        - items
    UpsertPageContentDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ContentItemDto'
      required:
        - items